refactor: phase 5 songs API handlers and shared emulator framework
