+++ src/apps/infinite-mac/components/InfiniteMacAppComponent.tsx
@@ -1,88 +1,16 @@
-import { useState } from "react";
+import { useMemo } from "react";
 import { AppProps } from "@/apps/base/types";
 import { AppWindowShell } from "@/components/shared/AppWindowShell";
 import { AppHelpAboutDialogs } from "@/components/shared/AppHelpAboutDialogs";
 import { appMetadata } from "../metadata";
-import { motion } from "motion/react";
 import { useInfiniteMacLogic } from "../hooks/useInfiniteMacLogic";
-import type { MacPreset } from "../hooks/useInfiniteMacLogic";
 import { InfiniteMacMenuBar } from "./InfiniteMacMenuBar";
 import { MAC_PRESETS } from "../hooks/useInfiniteMacLogic";
 import { PRESET_AVERAGE_COLORS } from "../presetAverageColors.generated";
+import { EmulatorPresetGrid } from "@/apps/shared-emulator/EmulatorPresetGrid";
 import { SquaresFour } from "@phosphor-icons/react";
 
 const FALLBACK_RGB = "169,175,190";
-
-function PresetGridCard({
-  preset,
-  onSelect,
-}: {
-  preset: MacPreset;
-  onSelect: () => void;
-}) {
-  const [thumbError, setThumbError] = useState(false);
-  const showThumb = !thumbError;
-  const textShadow = "0 1px 3px rgba(0,0,0,0.95)";
-  const rgb = PRESET_AVERAGE_COLORS[preset.id] ?? FALLBACK_RGB;
-  const bgColor = `rgb(${rgb})`;
-  const overlayColor = `rgba(${rgb},0.5)`;
-
-  return (
-    <motion.button
-      type="button"
-      onClick={onSelect}
-      className="group relative rounded overflow-hidden bg-neutral-800 hover:bg-neutral-700 transition-all duration-200 w-full flex flex-col min-h-[100px] @md:min-h-0 [box-shadow:0_4px_12px_rgba(0,0,0,0.5),0_0_0_1px_rgba(255,255,255,0.05)] hover:[box-shadow:0_8px_24px_rgba(0,0,0,0.6),0_0_0_1px_rgba(255,255,255,0.12)]"
-      whileTap={{
-        scale: 0.97,
-        y: 0,
-        transition: { type: "spring", duration: 0.15 },
-      }}
-    >
-      <div
-        className="w-full flex-1 min-h-0 relative shrink-0 overflow-hidden"
-        style={{
-          aspectRatio: `${preset.screenSize.width} / ${preset.screenSize.height}`,
-          backgroundColor: bgColor,
-        }}
-      >
-        {showThumb ? (
-          <img
-            src={preset.image}
-            alt=""
-            className="absolute inset-0 w-full h-full object-cover object-top opacity-80 transition-all duration-[800ms] ease-out group-hover:scale-105 group-hover:opacity-100"
-            onError={() => setThumbError(true)}
-          />
-        ) : null}
-        <div
-          className="absolute inset-0 pointer-events-none opacity-0 transition-opacity duration-200 group-hover:opacity-20"
-          style={{ backgroundColor: overlayColor }}
-          aria-hidden
-        />
-      </div>
-      <div
-        className="absolute inset-0 pointer-events-none z-[1]"
-        style={{
-          background: `linear-gradient(to top, ${bgColor} 0%, transparent 55%)`,
-        }}
-        aria-hidden
-      />
-      <div className="absolute bottom-0 left-2 right-2 pt-2 pb-2 flex flex-col items-start gap-0.5 @md:flex-row @md:justify-between @md:items-baseline z-10 pointer-events-none">
-        <span
-          className="text-white font-apple-garamond !text-[18px] leading-tight truncate max-w-full"
-          style={{ textShadow }}
-        >
-          {preset.name}
-        </span>
-        <span
-          className="text-neutral-300 text-[10px] shrink-0 opacity-100 @md:opacity-0 transition-opacity duration-200 @md:group-hover:opacity-100"
-          style={{ textShadow }}
-        >
-          {preset.year}
-        </span>
-      </div>
-    </motion.button>
-  );
-}
 
 export function InfiniteMacAppComponent({
   isWindowOpen,
@@ -114,6 +42,19 @@ export function InfiniteMacAppComponent({
     handleCaptureScreenshot,
     handleIframeLoad,
   } = useInfiniteMacLogic({ isWindowOpen, instanceId });
+
+  const presetCards = useMemo(
+    () =>
+      MAC_PRESETS.map((preset) => ({
+        id: preset.id,
+        name: preset.name,
+        year: preset.year,
+        image: preset.image,
+        rgb: PRESET_AVERAGE_COLORS[preset.id] ?? FALLBACK_RGB,
+        screenSize: preset.screenSize,
+      })),
+    []
+  );
 
   const menuBar = (
     <InfiniteMacMenuBar
@@ -232,15 +173,14 @@ export function InfiniteMacAppComponent({
                 </div>
 
                 <div className="flex-1 min-h-0 overflow-y-auto flex justify-start @md:justify-center w-full p-4 @container">
-                  <div className="preset-grid grid grid-cols-1 @md:grid-cols-3 gap-2 w-full max-w-md @md:max-w-none pb-[calc(1rem+env(safe-area-inset-bottom,0px))] @md:pb-0">
-                    {MAC_PRESETS.map((preset) => (
-                      <PresetGridCard
-                        key={preset.id}
-                        preset={preset}
-                        onSelect={() => handleSelectPreset(preset)}
-                      />
-                    ))}
-                  </div>
+                  <EmulatorPresetGrid
+                    presets={presetCards}
+                    layout="aspect-ratio"
+                    onSelectPreset={(presetId) => {
+                      const preset = MAC_PRESETS.find((p) => p.id === presetId);
+                      if (preset) handleSelectPreset(preset);
+                    }}
+                  />
                 </div>
               </div>
             )}
